Introduce fully lexically scoped variables
Adds a new notion of block scoped variables, declared using new syntax
"let x =".
Adds new block scope with "begin" and "end" keywords.
Existing block structures like while/if-elseif-else/try/for are also
lexically scopable.
Also adds explicit global variable assignment / declaration via
"global"
By default variables are global, to keep backwards compatibility with
LambdaMOO.

+++ crates/kernel/src/builtins/bf_server.rs
@@ -627,7 +627,7 @@ fn bf_call_function(bf_args: &mut BfCallState<'_>) -> Result<BfRet, BfErr> {
 
     // Then ask the scheduler to run the function as a continuation of what we're doing now.
     Ok(VmInstr(ExecutionResult::ContinueBuiltin {
-        bf_func_num: func_offset,
+        bf_func_num: func_offset as u16,
         arguments: args[..].to_vec(),
     }))
 }
